crypto draft V2
