+++ cryptocurrency/Dockerfile
@@ -1,0 +1,4 @@
+FROM ghcr.io/deephaven/grpc-api
+COPY cryptoApp.app /
+COPY crypto.py /
+run pip3 install pycoingecko
