+++ cryptocurrency/cryptoApp.app
@@ -1,0 +1,6 @@
+type=script
+scriptType=python
+enabled=true
+id=crypto
+name=Crypto application
+file_0=./crypto.py
