+++ cryptocurrency/start.sh
@@ -1,0 +1,2 @@
+docker build --tag crypto-deephaven/grpc-api .
+docker-compose up
